fixed duplicate table generation everytime connection is established
